# Remove unsafe char[] access in PlatformDependent