--- a/codec/src/main/java/io/netty/handler/codec/HeadersUtils.java
+++ b/codec/src/main/java/io/netty/handler/codec/HeadersUtils.java
@@ -211,7 +211,7 @@ public final class HeadersUtils {
 
         @Override
         public Object[] toArray() {
-            Object[] arr = new String[size()];
+            Object[] arr = new Object[size()];
             fillArray(arr);
             return arr;
         }