@Scottmitch 2016-06-29
```
With this PR
Benchmark                                          (size)   Mode  Cnt          Score         Error  Units
AsciiStringBenchmark.hashCodeBenchCharSequenceNew       3  thrpt    5  201066167.422 ± 8848348.606  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew       5  thrpt    5  161667082.187 ± 4461054.701  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew       7  thrpt    5  127944592.934 ± 5840062.089  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew       8  thrpt    5   66785310.236 ± 1365469.591  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew      10  thrpt    5   64217697.840 ± 1672072.188  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew      20  thrpt    5   52446059.707 ± 1469696.940  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew      50  thrpt    5   34614991.869 ±  632616.742  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew     100  thrpt    5   21969769.942 ±  637611.843  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew    1000  thrpt    5    2914633.221 ±   80374.365  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld       3  thrpt    5  169523487.095 ± 2813888.634  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld       5  thrpt    5  133310533.270 ± 1361342.811  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld       7  thrpt    5  107573994.906 ± 7157047.592  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld       8  thrpt    5  101098109.425 ± 2667872.262  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld      10  thrpt    5   96656523.706 ± 2373692.530  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld      20  thrpt    5   55012284.429 ± 1287338.167  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld      50  thrpt    5   25934497.756 ±  481047.217  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld     100  thrpt    5   12836197.598 ±  352280.212  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld    1000  thrpt    5    1271387.921 ±   26046.327  ops/s

Before PR

Benchmark                                          (size)   Mode  Cnt          Score          Error  Units
AsciiStringBenchmark.hashCodeBenchCharSequenceNew       3  thrpt    5  179524182.749 ±  6888675.061  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew       5  thrpt    5  155623444.126 ± 10712300.264  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew       7  thrpt    5  141975274.134 ±  9415597.369  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew       8  thrpt    5  106533137.232 ±  2806264.403  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew      10  thrpt    5   95156374.592 ±  3011730.087  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew      20  thrpt    5   71586258.970 ±  1290335.794  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew      50  thrpt    5   37653402.598 ±  1249025.500  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew     100  thrpt    5   22529631.282 ±   604240.749  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew    1000  thrpt    5    2523269.080 ±    44337.349  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld       3  thrpt    5  164688643.527 ±  6192564.871  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld       5  thrpt    5  129693577.896 ±  7092971.989  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld       7  thrpt    5  108274574.172 ±  6577438.748  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld       8  thrpt    5  101832213.389 ±  2129438.327  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld      10  thrpt    5   97074885.606 ±  4344009.584  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld      20  thrpt    5   60843708.034 ±  1259821.842  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld      50  thrpt    5   25151374.465 ±  1299056.138  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld     100  thrpt    5   12502531.315 ±   671122.104  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld    1000  thrpt    5    1267889.337 ±    31158.960  ops/s
```