@Scottmitch 2016-06-29
Here are results from `HashCodeCollisionTest`. I compared the new algorithm to the algorithm prior to any optimizations ... the amount of hash collisions is the same. 

```
==Old Duplicates==
Number duplicates: 3
2 duplicates. -1135485755: [NONLANGUAGE] [UPSWELLED] 
2 duplicates. 19007732: [EXCELLENCY] [TRACT] 
2 duplicates. 1939445677: [EULOGIUM] [OVERHARVESTING] 

==New Duplicates==
Number duplicates: 3
2 duplicates. -1847989755: [BANYANS] [PREMANUFACTURED] 
2 duplicates. 448129169: [HOMOGAMOUS] [IRREVOCABILITY] 
2 duplicates. -364122137: [MISMAKES] [MISTAKEN] 
```