@Scottmitch 2016-06-29
I also did some investigation as to how this hash code differs from the existing one in terms of which header names go into which bucket in `DefaultHeaders` I have these results but they are similar. The new algorithm may have a small advantage because some "frequently used headers" (`:path` and `:host`) no longer end up in the same bucket. However it may be a good followup exercise to see if we can get better distribution into the `DefaultHeader` buckets.