@normanmaurer 2016-06-29
@Scottmitch please fix check style :)