@buchgr 2016-06-29
@Scottmitch yay to deleting lots of code in `PlatformDependent`! Instead of checking in the whole word file, just read from `/usr/share/dict/words` ... making this more light-weight :).