@normanmaurer 2016-06-29
@Scottmitch we can do what @buchgr suggested as well :)