@Scottmitch 2016-06-29
I think there is value in using both sources for words (more data the better), but I agree it throws the commit statistics out of whack and also bloats the repo. I created a gist for the file https://gist.github.com/Scottmitch/de2f03912778016ecee3c140478f07e0#file-englishwords-txt so it can be used for any future hash code changes.