@carl-mastrangelo 2016-06-29
@Scottmitch I looked over this, but am afraid I am not as familiar with this.  Specifically, why is the 0x1f used pervasively?  I see the comment explaining it, but still isn't clear why it is necessary.  Also, if the CharSequence is a String, doesn't the hashCode need to be specifically the one defined for string?  Won't masking the bits make the two hashCodes differ?