@Scottmitch 2016-06-29
> Specifically, why is the 0x1f used pervasively? 

An interesting aspect of http headers is that they are compared in a case-insensitive manner. This is accomplished by supporting a equals/hashcode which is case-insensitive. `0x1f` masks out the bits which differ between upper case and lower case characters and thus provides a case-insensitive hash code.  Since many HTTP/2 semantics are shared with HTTP/1.x we also re-use some of the header definitions from codec-http1, share the hashCode (which is cached by `AsciiString`) algorithm, and use case-sensitive equals.