@Scottmitch 2016-06-29
> Also, if the CharSequence is a String, doesn't the hashCode need to be specifically the one defined for string?

Since we use `CharSequence` as a key, we want to make sure that content for different types of `CharSequence`s are compared consistently. For example the following behavior is generally expected:

```java
DefaultHttp2Headers headers = new DefaultHttp2Headers();
AsciiString asciiName = new AsciiString("foo");
headers.add(asciiName, "goo");
assertTrue(headers.contains("foo")); // (1)
assertTrue(headers.contains(asciiName)); // (2)
```

If we used the object's equals/hashCode then `(1)` would fail. This was originally to support the existing behavior of codec-http, but it is convenient for HTTP/2 which shares many semantics as HTTP/1.x  (and header names are also [limited to ASCII in the HTTP/2 rfc](https://tools.ietf.org/html/rfc7540#section-8.1.2)).

> Won't masking the bits make the two hashCodes differ?

Yes it is not expected that this hashCode and String.hashCode() will be equal, but they don't have to be as long as we consistently use the same algorithm.